tools/logo_builder/requirements.txt
# NHL Logo Builder Dependencies

# HTTP requests for downloading NHL data and SVG logos
requests>=2.31.0

# Image processing and manipulation
Pillow>=10.0.0

# SVG to PNG rasterization
resvg-py>=0.2.6

# Perceptual color quantization (logo_quantize.py)
numpy>=1.24.0
scikit-image>=0.21.0
scipy>=1.10.0
